fix: 타입명 수정 #112
