+++ src/app/api/reserve/insertNewReservation.ts
@@ -1,0 +1,29 @@
+import { ReserveInfo } from '@/types/reserve';
+import { supabase } from '../supabase/supabase';
+import { PostgrestSingleResponse } from '@supabase/supabase-js';
+import camelcaseKeys from 'camelcase-keys';
+
+export const insertNewReservation = async (reserveInfo: ReserveInfo) => {
+  const { userId, classId, reservePrice, reserveQuantity, timeId } = reserveInfo;
+
+  const { data, error }: PostgrestSingleResponse<{ reserve_id: string }> = await supabase
+    .from('reserve')
+    .insert([
+      {
+        user_id: userId,
+        class_id: classId,
+        reserve_price: reservePrice,
+        reserve_quantity: reserveQuantity,
+        time_id: timeId
+      }
+    ])
+    .select('reserve_id')
+    .single();
+
+  if (error) {
+    console.error('insertNewReservation 오류 =>', error);
+    return;
+  }
+
+  return data.reserve_id;
+};
